+++ examples/deps.ts
@@ -1,0 +1,2 @@
+export * from "../mod.ts";
+export { sleep } from "../runners/utility.ts";