+++ runners/utility.ts
@@ -105,3 +105,13 @@ export function randomString(length: number): string {
 
   return result.join("");
 }
+
+/**
+ * Sleep for a while.
+ * @param delayms Delay in milliseconds.
+ */
+export async function sleep(delayms: number): Promise<void> {
+  await new Promise<void>((resolve, _reject) =>
+    setTimeout(() => resolve(), delayms)
+  );
+}